refactor(api): move scorecard service
